backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-smorest==0.42.1
python-dotenv==1.0.0
marshmallow==3.20.1
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-cov==4.1.0
flake8==6.1.0
